css/html 変更
岩井 6/27
css/html 変更

+++ Cinema_e-grid/src/main/java/com/cinemaEgrid/web/Delete/UserAdminDeleteController.java
@@ -41,15 +41,15 @@ public class UserAdminDeleteController {
 	@RequestMapping(value = "/success", method = RequestMethod.GET)
 	private ModelAndView index2(@RequestParam("ID") String id, User form,
 			ModelAndView mav, Model model) {
-//		if (form.getUser_del_flg() == "1") {
+		if (form.getUser_del_flg() == "1") {
 			try {
 				UserAdminDao.deleteAdminUser(id);
 			} catch (SQLException e) {
 			}
 			mav.setViewName("Admin/Done/deleteDone");
-//		} else {
-//			mav.setViewName("redirect:/admin/alldisplay");
-//		}
+		} else {
+			mav.setViewName("redirect:/admin/alldisplay");
+		}
 		return mav;
 	}
 }
